+++ js/api/constants.js
@@ -5,3 +5,5 @@ export const apiBaseUrl = "https://api.noroff.dev";
 export const postsEndpoint = "/api/v1/social/posts/";
 
 export const login = "/api/v1/social/auth/login";
+
+export const register = "/api/v1/social/auth/register/";
